Widen slab_size from u32 to u64 to support slabs >= 4 GiB
The u32 type capped slab size at ~4 GiB which is an undesired
limitation. Updates the on-disk superblock and checkpoint formats
to serialize slab_size as 8 bytes. Also documents all FormatParams
fields with their units (bytes) and constraints.

+++ apps/extent-benchmark/src/main.rs
@@ -118,7 +118,7 @@ fn main() {
     let params = FormatParams {
         slab_size: config.slab_size,
         max_element_size: config.size_class,
-        metadata_block_size: config.slab_size.min(131072),
+        metadata_block_size: config.slab_size.min(131072) as u32,
         sector_size: 4096,
         region_count: 32,
     };
@@ -326,7 +326,7 @@ fn validate_config(config: &BenchmarkConfig) -> Result<(), String> {
         ));
     }
     if let Some(total) = config.total_size {
-        if total <= config.slab_size as u64 {
+        if total <= config.slab_size {
             return Err(format!(
                 "total-size ({total}) must be greater than slab-size ({})",
                 config.slab_size
